requirements.txt
# Core dependencies
networkx>=3.0
numpy>=1.24.0
matplotlib>=3.7.0
scikit-learn>=1.3.0

# Sentence embeddings
sentence-transformers>=2.2.0

# LLM APIs
anthropic>=0.18.0
openai>=1.10.0
google-genai>=0.3.0
python-dotenv>=1.0.0

# Optional utilities
pandas>=2.0.0
tqdm>=4.65.0
